:sparkles: feat(vm): 实现闭包操作与 Arc 引用计数指令
闭包操作：
- MakeClosure: 创建带环境的闭包，从 closure_env 收集 upvalue
- LoadUpvalue: 从闭包环境加载变量
- StoreUpvalue: 存储到闭包环境
- CloseUpvalue: 关闭并清理 upvalue

Arc 引用计数：
- ArcNew: 创建 Arc 包装
- ArcClone: 增加引用计数
- ArcDrop: 减少引用计数

其他：
- Cast: Int↔Float 类型转换

+++ src/vm/executor.rs
@@ -1532,7 +1532,8 @@ impl VM {
 
                 // 确保 closure_env 有足够的空间
                 if upvalue_idx as usize >= self.closure_env.len() {
-                    self.closure_env.resize(upvalue_idx as usize + 1, RuntimeValue::Unit);
+                    self.closure_env
+                        .resize(upvalue_idx as usize + 1, RuntimeValue::Unit);
                 }
                 self.closure_env[upvalue_idx as usize] = value;
             }
@@ -1821,7 +1822,7 @@ impl VM {
                     3 => {
                         // I32 → I64
                         if let RuntimeValue::Int(n) = self.regs.read(src) {
-                            self.regs.write(dst, RuntimeValue::Int(*n as i64));
+                            self.regs.write(dst, RuntimeValue::Int(*n));
                         } else {
                             return Err(VMError::TypeError("int".into()));
                         }
@@ -1837,7 +1838,7 @@ impl VM {
                     5 => {
                         // F32 → F64
                         if let RuntimeValue::Float(f) = self.regs.read(src) {
-                            self.regs.write(dst, RuntimeValue::Float(*f as f64));
+                            self.regs.write(dst, RuntimeValue::Float(*f));
                         } else {
                             return Err(VMError::TypeError("float".into()));
                         }
